1.添加菜谱功能（完善版） 2.完成登录和注册功能

+++ src/main/java/com/hh/recipe/utils/JwtHelper.java
@@ -32,12 +32,12 @@ public class JwtHelper {
     }
 
     //从token字符串获取userid
-    public Long getUserId(String token) {
+    public Integer getUserId(String token) {
         if (StringUtils.isEmpty(token)) return null;
         Jws<Claims> claimsJws = Jwts.parser().setSigningKey(tokenSignKey).parseClaimsJws(token);
         Claims claims = claimsJws.getBody();
         Integer userId = (Integer) claims.get("userId");
-        return userId.longValue();
+        return userId;
     }
